refactored update profile
