+++ src/types/enumTypes.ts
@@ -23,6 +23,8 @@ export enum expiryBadgeEnum {
 export enum accountTypeEnum {
   RENTER = "renter",
   MANAGER = "manager",
+  /** This was created for deleting account - when user has to log in again due to timeout */
+  GHOST = "ghost",
 }
 
 export enum confirmModalTypeEnum {
